+++ apps/nextjs-example/src/components/ReactQueryClientProvider.tsx
@@ -1,14 +1,12 @@
-"use client"
+"use client";
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { PropsWithChildren } from 'react';
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { PropsWithChildren } from "react";
 
 const queryClient = new QueryClient();
 
 export function ReactQueryClientProvider({ children }: PropsWithChildren) {
   return (
-    <QueryClientProvider client={queryClient}>
-      {children}
-    </QueryClientProvider>
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
   );
 }