+++ packages/wallet-adapter-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aptos-labs/wallet-adapter-react
 
+## 3.7.11
+
+### Patch Changes
+
+- Updated dependencies [b8cae2d]
+  - @aptos-labs/wallet-adapter-core@4.24.0
+
 ## 3.7.10
 
 ### Patch Changes